ZBUG-1315: An issue with data type range values for Mysql database tables

+++ store/src/java/com/zimbra/cs/db/DbVolumeBlobs.java
@@ -59,7 +59,7 @@ public final class DbVolumeBlobs {
             stmt.setShort(pos++, vol.getId());
             stmt.setInt(pos++, mbox.getId());
             stmt.setInt(pos++, item.getId());
-            stmt.setInt(pos++, item.getSavedSequence());
+            stmt.setLong(pos++, item.getSavedSequence());
             stmt.setString(pos++, item.getDigest());
             stmt.setBoolean(pos++, false);
             stmt.executeUpdate();
@@ -82,7 +82,7 @@ public final class DbVolumeBlobs {
             stmt.setShort(pos++, volId);
             stmt.setInt(pos++, info.mailboxId);
             stmt.setInt(pos++, info.itemId);
-            stmt.setInt(pos++, info.revision);
+            stmt.setLong(pos++, info.revision);
             stmt.setString(pos++, info.digest);
             stmt.setBoolean(pos++, false);
             stmt.executeUpdate();
